requirements.txt
openai>=1.55.0
python-dotenv>=1.0.1
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
requests>=2.31.0

